requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
rich>=13.0.0

# Embedding and vector storage
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA support

# Language model interfaces
openai>=1.0.0
transformers>=4.30.0
langchain>=0.1.0
langchain-community>=0.0.10

# Code parsing and analysis
tree-sitter>=0.20.1
pygments>=2.15.0
gitpython>=3.1.30

# Web server
websockets>=11.0.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
yaspin>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0